double para float real0m1.266s user0m4.288s sys0m0.024s

+++ lib/imagem.h
@@ -1,1 +1,1 @@
-void gera_imagem(double**, int, int);
+void gera_imagem(float**, int, int);
